requirements/common.txt
selenium==4.15.2
Appium-Python-Client==3.1.0
allure-pytest==2.12.0
dependency-injector-fork==4.42.1
jira==3.4.1
requests==2.28.1
Pillow==10.3.0
pytest==7.4.0
python-dotenv==0.21.0
pytest-testrail==2.9.0
pytest-random-order==1.1.0
phonenumbers==8.13.35
typing-extensions==4.10.0
testrail_api==1.13.0
schema==0.7.5
google-api-python-client==2.37.0
pyjwt==2.8.0